% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytest-package.R
\docType{package}
\name{lazytest-package}
\alias{lazytest}
\alias{lazytest-package}
\title{lazytest: Runs Only Failing Tests}
\description{
A companion package to the testthat package that saves development time by helping you rerun only the tests that have failed during the last run.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cynkra/lazytest}
  \item \url{https://cynkra.github.io/lazytest/}
  \item Report bugs at \url{https://github.com/cynkra/lazytest/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

}
\keyword{internal}
