% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytest_local.R
\name{lazytest_local}
\alias{lazytest_local}
\title{Test a local source package}
\usage{
lazytest_local(
  path = ".",
  reporter = NULL,
  ...,
  lazytest_reset = FALSE,
  stop_on_failure = TRUE,
  stop_on_warning = FALSE,
  filter = NULL,
  load_package = "source"
)
}
\arguments{
\item{path}{Path to directory containing tests.}

\item{reporter}{Reporter to use to summarise output. Can be supplied
as a string (e.g. "summary") or as an R6 object
(e.g. \code{SummaryReporter$new()}).

See \link[testthat]{Reporter} for more details and a list of built-in reporters.}

\item{...}{Additional arguments passed to \code{\link[testthat:test_dir]{test_dir()}}}

\item{lazytest_reset}{Set to \code{TRUE} to run all tests, regardless of what the
last test results were.}

\item{filter}{Must be \code{NULL}.}

\item{load_package}{Strategy to use for load package code:
\itemize{
\item "none", the default, doesn't load the package.
\item "installed", uses \code{\link[=library]{library()}} to load an installed package.
\item "source", uses \code{\link[pkgload:load_all]{pkgload::load_all()}} to a source package.
To configure the arguments passed to \code{load_all()}, add this
field in your DESCRIPTION file:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Config/testthat/load-all: list(export_all = FALSE, helpers = FALSE)
}\if{html}{\out{</div>}}
}}
}
\description{
A drop-in replacement for \code{testthat::test_local()} that by default reruns
only the tests that failed last time.
}
\examples{
\dontrun{
lazytest_local()
}
}
